ev_charging_recommender_ai/requirements.txt
flask
requests
scikit-learn
geopy
python-dotenv
joblib
gunicorn
